rename style sheets
