+++ src/components/FormListModal.vue
@@ -92,7 +92,7 @@
 <script lang="ts">
 // import 'bootstrap/dist/css/bootstrap.min.css'
 // import 'bootstrap-vue/dist/bootstrap-vue.css'
-import '../camundaFormIOFormList.scss'
+import '../styles/camundaFormIOFormList.scss'
 import { Component, Prop, Vue } from 'vue-property-decorator'
 import CamundaRest from '../services/camunda-rest'
 import { Form } from 'vue-formio';
