+++ src/services/utils.ts
@@ -9,7 +9,7 @@ export const findFilterKeyOfAllTask = (array: string|any[], key: string|number, 
       return array[i]["id"];
     }
   }
-  return null;
+  return array[0]["id"];
 }
 
 export const TASK_FILTER_LIST_DEFAULT_PARAM =  [{sortBy: "created",
