+++ src/styles/camundaFormIOTaslistSearch.scss
